pyproject.toml
[project]
name = "tavily-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-mcp-adapters>=0.1.9",
    "mcp[cli]>=1.12.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
